Added zoom ad basic colour offset

+++ tfrewin.play.fractal.start/processor/IFormulaProcessor.cs
@@ -5,6 +5,6 @@ namespace tfrewin.play.fractal.start.processor
 {
     public interface IFormulaProcesor
     {
-        Matrix Process(int planeWidth, int planeHeight, int zoom, int maximumIteration);
+        Matrix Process(int planeWidth, int planeHeight, double zoom, int maximumIteration);
     }
 }
